move helper function from init to utils
